requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
asttokens==2.2.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.0.1
beautifulsoup4==4.12.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
chroma-hnswlib==0.7.2
chromadb==0.4.5
click==8.1.6
colorama==0.4.6
coloredlogs==15.0.1
decorator==5.1.1
elevenlabs==0.2.24
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.99.1
filelock==3.12.2
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.6.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-resources==6.0.1
ipython==8.14.0
jedi==0.19.0
Jinja2==3.1.2
lxml==4.9.3
MarkupSafe==2.1.3
matplotlib-inline==0.1.6
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
networkx==3.1
numpy==1.25.2
onnxruntime==1.15.1
openai==0.27.8
outcome==1.2.0
overrides==7.4.0
packaging==23.1
parso==0.8.3
pickleshare==0.7.5
posthog==3.0.1
prompt-toolkit==3.0.39
protobuf==4.24.0
pulsar-client==3.2.0
pure-eval==0.2.2
pycparser==2.21
pydantic==1.10.7
Pygments==2.16.1
PyPDF2==3.0.1
PyPika==0.48.9
pyreadline3==3.4.1
PySocks==1.7.1
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
starlette==0.27.0
sympy==1.12
tenacity==8.2.2
termcolor==2.3.0
tokenizers==0.13.3
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.0.0+cpu
tqdm==4.65.0
traitlets==5.9.0
transformers==4.31.0
trio==0.22.2
trio-websocket==0.10.3
typing_extensions==4.7.1
urllib3==2.0.4
uvicorn==0.23.2
watchfiles==0.19.0
wcwidth==0.2.6
webdriver-manager==4.0.0
websockets==11.0.3
wsproto==1.2.0
yarl==1.9.2
werkzeug==2.3.6
python-multipart==0.0.6